chore: refactor GDD edit GUI
