--- apps/app/src/react/components/sidebar/timelineObj/GDD/lib.ts
@@ -1,3 +1,0 @@
-export function getBasicType(schemaType: string | string[]): string {
-	return Array.isArray(schemaType) ? schemaType[0] : schemaType
-}
